requirements.txt
Flask
pandas
google-api-python-client
python-dotenv
emoji
nltk
transformers
torch
